refactor(checker): split generator declaration yield recovery
