+++ crates/tsz-checker/src/types/function_type.rs
@@ -1,6 +1,7 @@
 //! Function, method, and arrow function type resolution.
 mod contextual_arity;
 mod function_name_diagnostics;
+mod generator_declaration_yield;
 mod js_prototype;
 mod jsx_body_context;
 mod literal_context;
@@ -15,7 +16,6 @@ use crate::query_boundaries::common::ContextualTypeContext;
 use crate::query_boundaries::type_checking_utilities as type_query;
 use crate::state::CheckerState;
 use tsz_parser::parser::NodeIndex;
-use tsz_parser::parser::node::NodeAccess;
 use tsz_parser::parser::syntax_kind_ext;
 use tsz_scanner::SyntaxKind;
 use tsz_solver::{TypeId, TypeParamInfo, TypeParamOrigin};
@@ -23,41 +23,6 @@ impl<'a> CheckerState<'a> {
     /// Get type of function declaration/expression/arrow.
     pub(crate) fn get_type_of_function(&mut self, idx: NodeIndex) -> TypeId {
         self.get_type_of_function_impl(idx, &TypingRequest::NONE)
-    }
-
-    /// Whether `node_idx` (a generator body subtree) contains a `yield` /
-    /// `yield*` expression belonging to *this* generator — i.e. not one nested
-    /// inside another function/class. Used to tell a genuinely empty generator
-    /// (`function* g() {}` → `Generator<never>`) apart from one whose only
-    /// `yield* expr` delegate could not be resolved during signature inference.
-    fn generator_body_contains_yield(&self, node_idx: NodeIndex, is_root: bool) -> bool {
-        let Some(node) = self.ctx.arena.get(node_idx) else {
-            return false;
-        };
-        if !is_root
-            && matches!(
-                node.kind,
-                k if k == syntax_kind_ext::FUNCTION_DECLARATION
-                    || k == syntax_kind_ext::FUNCTION_EXPRESSION
-                    || k == syntax_kind_ext::METHOD_DECLARATION
-                    || k == syntax_kind_ext::GET_ACCESSOR
-                    || k == syntax_kind_ext::SET_ACCESSOR
-                    || k == syntax_kind_ext::CONSTRUCTOR
-                    || k == syntax_kind_ext::CLASS_DECLARATION
-                    || k == syntax_kind_ext::CLASS_EXPRESSION
-            )
-        {
-            return false;
-        }
-        if node.kind == syntax_kind_ext::YIELD_EXPRESSION {
-            return true;
-        }
-        for child in self.ctx.arena.get_children(node_idx) {
-            if self.generator_body_contains_yield(child, false) {
-                return true;
-            }
-        }
-        false
     }
 
     pub(crate) fn get_type_of_function_impl(
@@ -71,23 +36,14 @@ impl<'a> CheckerState<'a> {
         let Some(node) = self.ctx.arena.get(idx) else {
             return TypeId::ERROR; // Missing node - propagate error
         };
-        // Determine if this is a function expression or arrow function (a closure)
         let is_closure = matches!(
             node.kind,
             syntax_kind_ext::FUNCTION_EXPRESSION | syntax_kind_ext::ARROW_FUNCTION
         );
-        // Track object-literal method shorthand for implicit-any like an
-        // arrow/function-expression member; `is_closure` (which also drives
-        // narrowing depth) stays arrow/function-expression only. Rationale and
-        // class-method exclusion: see `is_object_literal_method`.
         let tracks_implicit_any = is_closure || self.is_object_literal_method(idx);
-        // Rule #42: Increment closure depth when entering a function expression or arrow function
-        // This causes mutable variables (let/var) to lose narrowing inside the closure
         if is_closure {
             self.ctx.inside_closure_depth += 1;
         }
-        // Helper macro to decrement closure depth before returning
-        // This ensures we properly track closure depth even on early returns
         macro_rules! return_with_cleanup {
             ($expr:expr) => {{
                 if is_closure {
@@ -150,7 +106,6 @@ impl<'a> CheckerState<'a> {
                 (false, false)
             };
 
-        // Function declarations don't report implicit any for parameters (handled by check_statement)
         let is_function_declaration = node.kind == syntax_kind_ext::FUNCTION_DECLARATION;
         let is_method_or_constructor = matches!(
             node.kind,
@@ -158,12 +113,9 @@ impl<'a> CheckerState<'a> {
         );
         let is_arrow_function = node.kind == syntax_kind_ext::ARROW_FUNCTION;
 
-        // Check for duplicate parameter names in function expressions and arrow functions (TS2300)
         if !is_function_declaration && !is_method_or_constructor {
-            // Check for required parameters following optional parameters (TS1016)
             self.check_parameter_ordering(parameters, Some(idx));
             self.check_binding_pattern_optionality(&parameters.nodes, body.is_some(), Some(idx));
-            // Check that rest parameters have array types (TS2370)
             self.check_rest_parameter_types(&parameters.nodes);
             self.check_strict_mode_reserved_parameter_names(
                 &parameters.nodes,
@@ -179,7 +131,6 @@ impl<'a> CheckerState<'a> {
             function_is_generator,
         );
 
-        // Push enclosing type parameters so nested functions can reference outer generic scopes.
         let enclosing_type_param_updates = self.push_enclosing_type_parameters(idx);
 
         self.exclude_params_for_type_param_constraints(parameters);
@@ -191,12 +142,10 @@ impl<'a> CheckerState<'a> {
             self.check_type_parameters_for_missing_names(type_parameters);
         }
 
-        // Check for unused type parameters (TS6133) in function expressions and arrows
         if !is_function_declaration && !is_method_or_constructor {
             self.check_unused_type_params(type_parameters, idx);
         }
 
-        // Collect parameter info using solver's ParamInfo struct
         let mut params = Vec::new();
         let mut param_types: Vec<Option<TypeId>> = Vec::new();
         let mut destructuring_context_param_types: Vec<Option<TypeId>> = Vec::new();
@@ -1960,81 +1909,22 @@ impl<'a> CheckerState<'a> {
                 self.ctx.had_outer_loop = saved_cf_context.3;
             }
 
-            // Generator function *declarations* defer their body walk to
-            // `check_function_declaration` (which maintains the full
-            // type-parameter scope chain), so the inline yield collection above
-            // is skipped for them and `final_generator_yield_type` stays unset.
-            // The synthesized signature would then be `Generator<any, …>` /
-            // `AsyncGenerator<any, …>`, dropping the inferred yield type for both
-            // `.d.ts` emit *and* call-site checking (e.g. a missed TS2322 on
-            // `g().next().value`). Methods and function expressions never hit
-            // this because they run the body walk inline above. Recover the
-            // inferred yield here with a diagnostics-suppressed body pass that
-            // reuses the same collection + `check_generator_body_return` the
-            // non-declaration path uses; the real TS7055/TS7025 and body
-            // diagnostics still come from `check_function_declaration`.
             if is_function_declaration
                 && is_generator
                 && !has_type_annotation
                 && final_generator_yield_type.is_none()
             {
-                let yield_diag_snapshot = DiagnosticSpeculationSnapshot::new(&self.ctx);
-                let saved_cf_context = (
-                    self.ctx.iteration_depth,
-                    self.ctx.switch_depth,
-                    self.ctx.label_stack.len(),
-                    self.ctx.had_outer_loop,
+                final_generator_yield_type = self.infer_generator_declaration_yield_type(
+                    body,
+                    contextual_type,
+                    has_type_annotation,
+                    annotated_return_type,
+                    return_type,
+                    type_annotation,
+                    idx,
+                    function_is_async,
+                    early_yield_type,
                 );
-                self.ctx.iteration_depth = 0;
-                self.ctx.switch_depth = 0;
-                let saved_yield_collection =
-                    std::mem::take(&mut self.ctx.generator_yield_operand_types);
-                let saved_had_ts7057 = std::mem::replace(&mut self.ctx.generator_had_ts7057, false);
-
-                let body_request = self.function_body_statement_request(false, contextual_type);
-                self.check_function_body_statement_with_own_literal_context(body, &body_request);
-                let inferred_yield =
-                    self.check_generator_body_return(GeneratorBodyReturnCheckCtx {
-                        is_generator,
-                        has_type_annotation,
-                        annotated_return_type,
-                        return_type,
-                        type_annotation,
-                        idx,
-                        function_is_async,
-                        early_yield_type,
-                        name_node: None,
-                        name_for_error: None,
-                    });
-                // Adopt the inferred yield only when it is concrete, or when the
-                // body has no `yield` at all — a genuinely empty generator
-                // (`function* g() {}`) yields `never`, matching tsc. A `never`
-                // *with* yields present means a `yield* expr` delegate whose
-                // element type could not be resolved during signature inference
-                // (a separate, pre-existing gap that also affects generator
-                // methods/expressions); keep the prior `any` fallback there
-                // rather than emit a less-accurate `never`.
-                final_generator_yield_type = match inferred_yield {
-                    Some(yield_t)
-                        if yield_t != TypeId::NEVER
-                            || !self.generator_body_contains_yield(body, true) =>
-                    {
-                        Some(yield_t)
-                    }
-                    _ => None,
-                };
-
-                self.ctx.generator_yield_operand_types = saved_yield_collection;
-                self.ctx.generator_had_ts7057 = saved_had_ts7057;
-                self.ctx.iteration_depth = saved_cf_context.0;
-                self.ctx.switch_depth = saved_cf_context.1;
-                self.ctx.label_stack.truncate(saved_cf_context.2);
-                self.ctx.had_outer_loop = saved_cf_context.3;
-                // Drop any body expression types this suppressed pass cached so
-                // `check_function_declaration` re-checks the body from scratch
-                // with the full type-parameter scope chain.
-                self.invalidate_function_body_for_param_retyping(body);
-                yield_diag_snapshot.rollback(&mut self.ctx.diagnostic_state());
             }
             self.pop_return_type();
             self.ctx.pop_yield_type();
